+++ src/components/Hero/Hero.tsx
@@ -30,7 +30,10 @@ export const heroSecondaryTypes = [
   "secondaryWhatsOn",
 ];
 export interface HeroImageProps
-  extends Pick<ComponentImageProps, "alt" | "src"> {}
+  extends Pick<
+    ComponentImageProps,
+    "alt" | "fallbackSrc" | "src" | "onError"
+  > {}
 export interface HeroProps {
   /**
    * Optional background color for the backdrop only in the `campaign` variant.
@@ -55,10 +58,10 @@ export interface HeroProps {
   /** Used to control how the `Hero` component will be rendered. */
   heroType?: HeroTypes;
   /** Object used to create and render the `Image` component. Note that only
-   * `src` and `alt` are the available attributes to pass. If `imageProps.alt`
-   * is left blank, a warning will be logged to the console and will cause
-   * accessibility issues. For `imageProps.src`, it will only work for the
-   * "secondary", "fiftyFifty" and "campaign" `Hero` types; Note: `imageProps.src`
+   * `alt`, `component`, and `src` are the available attributes to pass. If
+   * `imageProps.alt` is left blank, a warning will be logged to the console and
+   * will cause accessibility issues. For `imageProps.src`, it will only work for
+   * the "secondary", "fiftyFifty" and "campaign" `Hero` types; Note: `imageProps.src`
    * can only be used in conjunction with `backgroundImageSrc` for the "campaign"
    * `Hero` type. Note: not all `Hero` variations utilize this prop. */
   imageProps?: HeroImageProps;
@@ -107,7 +110,8 @@ export const Hero: ChakraComponent<
 
       if (imageProps.src && !imageProps.alt) {
         console.warn(
-          `NYPL Reservoir Hero: The "imageProps.src" prop was passed but the "imageProps.alt" props was not. This will make the rendered image inaccessible.`
+          `NYPL Reservoir Hero: The "imageProps.src" prop was passed but the "imageProps.alt"` +
+            ` props was not. This will make the rendered image inaccessible.`
         );
       }
 
@@ -176,7 +180,7 @@ export const Hero: ChakraComponent<
        * into the component file and the related styles for all variants, other
        * than the "secondary" variant, were removed from the theme file. */
       const allDefaultBackgroundColors = {
-        primary: useColorModeValue("ui.bg.default", "dark.ui.bg.default"),
+        primary: useColorModeValue("ui.black", "dark.ui.bg.default"),
         secondary: useColorModeValue("ui.bg.default", "dark.ui.bg.default"),
         tertiary: useColorModeValue("ui.gray.x-dark", "dark.ui.bg.default"),
         campaign: useColorModeValue("dark.ui.bg.default", "dark.ui.bg.default"),
@@ -258,7 +262,7 @@ export const Hero: ChakraComponent<
           ...(foregroundColor && { color: foregroundColor }),
           ...(backgroundColor
             ? { backgroundColor }
-            : { defaultBackgroundColor }),
+            : { bgColor: defaultBackgroundColor }),
         };
       } else if (
         foregroundColor ||
@@ -272,6 +276,15 @@ export const Hero: ChakraComponent<
         );
       }
 
+      const imageToRender = (
+        <Image
+          alt={imageProps.alt}
+          fallbackSrc={imageProps.fallbackSrc}
+          onError={imageProps.onError}
+          src={imageProps.src}
+        />
+      );
+
       const childrenToRender =
         heroType === "campaign" ? (
           <>
@@ -281,7 +294,7 @@ export const Hero: ChakraComponent<
                 backgroundImage: `url(${imageProps.src})`,
               }}
             >
-              <Image alt={imageProps.alt} src={imageProps.src} />
+              {imageToRender}
             </Box>
             <Box __css={styles.interior}>
               {finalHeading}
@@ -300,7 +313,7 @@ export const Hero: ChakraComponent<
                       : undefined,
                 }}
               >
-                <Image alt={imageProps.alt} src={imageProps.src} />
+                {imageToRender}
               </Box>
             )}
             {finalHeading}