+++ CHANGELOG.md
@@ -8,8 +8,21 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Adds
+
+- Adds docs for `"Visited State Patterns"` to the `Link` component story page.
+- Adds the `ui.warning.tertiary` color token in order to have a light mode `"warning"` color option that meets the WCAG 2.1 AA color contrast standard.
+- Adds `fallbackSrc` and `onError` props to the `Image` component.
+- Adds the `fallbackSrc` and `onError` properties to the `imageProps` prop for the `Hero` component.
+
 ### Updates
 
+- Updates the placeholder value for the search field within the `MultiSelect` component.
+- Updates the background colors for the `"noBrand"` variant of the `Button` component.
+- Updates the hex values for the `ui.warning.primary` and `ui.warning.secondary` color tokens.
+- Updates the `Banner` and `StatusBadge` components to use the `ui.warning.tertiary` color for bertter accessibility in the `"warning"` variant.
+- Updates the styles for the UI colors, text treatment and `Clear all` button in the `TagSet` component.
+- Updates the component category for the `TagSet` component to `"Content Display"`.
 - Updates placeholder images on story pages and tests to use the `getPlaceholderImage` utility function which provides NYPL images.
 
 ## 3.1.0 (April 11, 2024)
@@ -31,6 +44,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Adds the `"editorMode"` option (pencil) to the `Icon` component.
 - Updates the `ul` bullet color in the `List` component.
 - Adds the `Errors` page to the `Accessibility Guide` section of Storybook.
+- Adds the `@storyook/addon-interactions` addon to Storybook and `@storybook/test` for writing interaction tests.
 
 ### Updates